Additional reorganization.
